Revert "ZCS-8785:Fix for issue with data type range values for Mysql"
This reverts commit 11f1edf7981c266328c19d7310efd4e9ca03a094.

+++ common/src/java/com/zimbra/common/mailbox/FolderStore.java
@@ -38,7 +38,7 @@ public interface FolderStore extends BaseFolderInfo {
     public boolean inTrash();
     public boolean isVisibleInImap(boolean displayMailFoldersOnly);
     public List<ACLGrant> getACLGrants();
-    public long getUIDValidity();
+    public int getUIDValidity();
     /** @return number of items in folder, including IMAP \Deleted item */
     public int getImapMessageCount();
     /** @return number of unread items in folder, including IMAP \Deleted items */
